Fix for a dangling pointer issue.
Previously values->getAsString would return a std::string variable, and
c_str a pointer to that variable, which would get assigned to argval. The
string object had no guaranteed lifetime in this context. GCC was keeping
the string in some kind of buffer so the error was not being caught. Clang
was reusing the memory location so argval[0] was changed to the value of
varval[1] when the loop continued. This fix insures the strings have a
lifetime long enough that psf-readprefs can copy them into its memory.